+++ tomcat/src/main/java/org/apache/catalina/servlet/Controller.java
@@ -5,5 +5,5 @@ import org.apache.coyote.http11.response.HttpResponse;
 
 public interface Controller {
 
-    void service(HttpRequest request, HttpResponse response) throws Exception;
+    void service(HttpRequest request, HttpResponse response);
 }
