+++ tomcat/src/test/java/nextstep/org/apache/coyote/http11/Http11ProcessorTest.java
@@ -2,6 +2,7 @@ package nextstep.org.apache.coyote.http11;
 
 import nextstep.jwp.servlet.ControllerAdvice;
 import nextstep.jwp.servlet.DispatcherServlet;
+import nextstep.jwp.servlet.RequestMapper;
 import org.apache.coyote.http11.Http11Processor;
 import org.junit.jupiter.api.Test;
 import support.StubSocket;
@@ -19,7 +20,7 @@ class Http11ProcessorTest {
     void process() {
         // given
         final var socket = new StubSocket();
-        final var processor = new Http11Processor(socket, new DispatcherServlet(new ControllerAdvice()));
+        final var processor = new Http11Processor(socket, new DispatcherServlet(new ControllerAdvice(), new RequestMapper()));
 
         // when
         processor.process(socket);
@@ -46,7 +47,7 @@ class Http11ProcessorTest {
                 "");
 
         final var socket = new StubSocket(httpRequest);
-        final Http11Processor processor = new Http11Processor(socket, new DispatcherServlet(new ControllerAdvice()));
+        final Http11Processor processor = new Http11Processor(socket, new DispatcherServlet(new ControllerAdvice(), new RequestMapper()));
 
         // when
         processor.process(socket);
